data_scraper/requirements.txt
beautifulsoup4==4.11.1
boto3==1.18.21
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
google-cloud-core==0.28.1
google-cloud-storage==1.10.0
lxml==4.9.2
requests==2.28.1
tqdm==4.64.1